# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Employee App",
        "title": "The Employee App",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/department/createDepartment": {
            "post": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Создание нового отдела.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Создание нового отдела.",
                "operationId": "createDepartment",
                "parameters": [
                    {
                        "description": "Данные нового отдела.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.EntityCreatingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/department/deleteDepartment/{id}": {
            "delete": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Удаление отдела.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Удаление отдела.",
                "operationId": "deleteDepartment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор отдела.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном удалении.",
                        "schema": {
                            "$ref": "#/definitions/transport.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/department/getAllDepartments": {
            "get": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Получение списка всех отделов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Получение списка всех отделов.",
                "operationId": "getAllDepartments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Department"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/department/getDepartment/{id}": {
            "get": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Получение информации об отделе.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Получение отдела.",
                "operationId": "getDepartment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор отдела.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/department/updateDepartment/{id}": {
            "put": {
                "security": [
                    {
                        "AuthorizationKey": []
                    }
                ],
                "description": "Обновление данных отдела.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "department"
                ],
                "summary": "Обновление данных отдела.",
                "operationId": "updateDepartment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор отдела.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные отдела.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Department"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном обновлении.",
                        "schema": {
                            "$ref": "#/definitions/transport.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/signIn": {
            "post": {
                "description": "Аутентификация пользователя (получение токена).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Аутентификация.",
                "operationId": "signIn",
                "parameters": [
                    {
                        "description": "Данные пользователя.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.SignInResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/signUp": {
            "post": {
                "description": "Регистрация пользователя в системе.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация.",
                "operationId": "signUp",
                "parameters": [
                    {
                        "description": "Данные нового пользователя.",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.EntityCreatingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/transport.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Department": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Идентификатор отдела.",
                    "type": "integer"
                },
                "name": {
                    "description": "Наименование отдела.",
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "Пароль пользователя.",
                    "type": "string"
                },
                "username": {
                    "description": "Имя поьзователя.",
                    "type": "string"
                }
            }
        },
        "transport.EntityCreatingResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Идентификатор сущности.",
                    "type": "integer"
                }
            }
        },
        "transport.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Сообщение с информацией об ошибке.",
                    "type": "string"
                }
            }
        },
        "transport.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Сообщение.",
                    "type": "string"
                }
            }
        },
...
        "transport.SignInResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "Токен авторизации.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AuthorizationKey": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}